first/pyproject.toml
[project]
name = "first"
version = "0.1.0"
description = "first using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.108.0,<1.0.0",
    "google-generativeai>=0.8.4",
]

[project.scripts]
kickoff = "first.main:kickoff"
plot = "first.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
